F-16C migration to Module
