+++ Scripts/DCS-BIOS/doc/Addresses.h
@@ -11930,12 +11930,12 @@
 #define F_16C_50_CMDS_01_EXP_CAT_SW_AM 0x445A, 0x4000
 #define F_16C_50_CMDS_02_EXP_CAT_SW 0x445A, 0x8000, 15
 #define F_16C_50_CMDS_02_EXP_CAT_SW_AM 0x445A, 0x8000
-#define F_16C_50_CMDS_CH_Amount_A 0x45A8
+#define F_16C_50_CMDS_CH_AMOUNT_A 0x45A8
 #define F_16C_50_CMDS_CH_EXP_CAT_SW 0x445E, 0x0001, 0
 #define F_16C_50_CMDS_CH_EXP_CAT_SW_AM 0x445E, 0x0001
 #define F_16C_50_CMDS_DISPENSE_BTN 0x445A, 0x0200, 9
 #define F_16C_50_CMDS_DISPENSE_BTN_AM 0x445A, 0x0200
-#define F_16C_50_CMDS_FL_Amount_A 0x45AC
+#define F_16C_50_CMDS_FL_AMOUNT_A 0x45AC
 #define F_16C_50_CMDS_FL_EXP_CAT_SW 0x445E, 0x0002, 1
 #define F_16C_50_CMDS_FL_EXP_CAT_SW_AM 0x445E, 0x0002
 #define F_16C_50_CMDS_JETT_SW 0x445A, 0x2000, 13
@@ -11945,8 +11945,8 @@
 #define F_16C_50_CMDS_MODE_KNB 0x445E, 0x00E0, 5
 #define F_16C_50_CMDS_MWS_SOURCHE_SW 0x445A, 0x1000, 12
 #define F_16C_50_CMDS_MWS_SOURCHE_SW_AM 0x445A, 0x1000
-#define F_16C_50_CMDS_O1_Amount_A 0x45A0
-#define F_16C_50_CMDS_O2_Amount_A 0x45A4
+#define F_16C_50_CMDS_O1_AMOUNT_A 0x45A0
+#define F_16C_50_CMDS_O2_AMOUNT_A 0x45A4
 #define F_16C_50_CMDS_PROG_KNB 0x445E, 0x001C, 2
 #define F_16C_50_CMDS_PWR_SOURCHE_SW 0x445A, 0x0400, 10
 #define F_16C_50_CMDS_PWR_SOURCHE_SW_AM 0x445A, 0x0400
